Move simple array accessors to a new file
